Move MockServer to its own file
