+++ apps/hops-fileshare/test/Helpers.kt
@@ -9,12 +9,8 @@ import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.testing.TestApplicationEngine
 import io.ktor.server.testing.withTestApplication
-import java.util.LinkedList
 import no.nav.security.mock.oauth2.MockOAuth2Server
-import okhttp3.mockwebserver.Dispatcher
 import okhttp3.mockwebserver.MockResponse
-import okhttp3.mockwebserver.MockWebServer
-import okhttp3.mockwebserver.RecordedRequest
 
 object Helpers {
     val oAuthMock = MockOAuth2Server()
@@ -142,46 +138,3 @@ private fun Application.config() = (environment.config as MapApplicationConfig).
     put("fileStore.virusScannerUrl", "${Helpers.virusScannerMockServer.getBaseUrl()}/scan")
 }
 
-class MockServer() {
-    val mockWebServer = MockWebServer()
-    private val dispatchChain = LinkedList<MatchAndDispatch>()
-
-    init {
-        mockWebServer.dispatcher = object : Dispatcher() {
-            override fun dispatch(request: RecordedRequest): MockResponse {
-                dispatchChain.forEach {
-                    if (it.matcher.invoke(request)) {
-                        return it.dispatcher.invoke(request)
-                    }
-                }
-                return defaultDispatcher(request)
-            }
-        }
-    }
-
-    var defaultDispatcher: (RecordedRequest) -> MockResponse = {
-        MockResponse()
-            .setResponseCode(HttpStatusCode.NotFound.value)
-    }
-
-    fun start() = mockWebServer.start()
-    fun shutdown() = mockWebServer.shutdown()
-
-    private class MatchAndDispatch(
-        val matcher: (RecordedRequest) -> Boolean,
-        val dispatcher: (RecordedRequest) -> MockResponse
-    )
-
-    fun getBaseUrl(): String = mockWebServer.url("/").toString().removeSuffix("/")
-
-    fun matchRequest(
-        matcher: (RecordedRequest) -> Boolean,
-        dispatcher: (RecordedRequest) -> MockResponse
-    ) {
-        dispatchChain.addFirst(MatchAndDispatch(matcher, dispatcher))
-    }
-
-    fun anyRequest(dispatcher: (RecordedRequest) -> MockResponse) {
-        dispatchChain.addFirst(MatchAndDispatch({ true }, dispatcher))
-    }
-}
